+++ modules/graphical.nix
@@ -33,7 +33,7 @@
     programs._1password.enable = true;
   };
 
-  hmModule = { pkgs, ... }: {
+  hmModule = { config, pkgs, ... }: {
     programs.vscode.package = pkgs.vscode;
 
     home.file.".ssh/config".text = let
@@ -49,5 +49,47 @@
       Host *
         ServerAliveInterval 120
     '';
+
+    programs.firefox.profiles.base.isDefault = false;
+
+    programs.firefox.profiles.personal = {
+      id = 1;
+
+      inherit (config.programs.firefox.profiles.base) search settings;
+
+      extensions = config.programs.firefox.profiles.base.extensions ++ [
+        pkgs.firefox-addons.bing-chat-for-all-browsers
+        pkgs.firefox-addons.copy-selected-links
+        pkgs.firefox-addons.ff2mpv
+        pkgs.firefox-addons.hover-zoom-plus
+        pkgs.firefox-addons.improved-tube
+        pkgs.firefox-addons.multi-account-containers
+        pkgs.firefox-addons.old-reddit-redirect
+        pkgs.firefox-addons.reddit-enhancement-suite
+        pkgs.firefox-addons.redirector
+        pkgs.firefox-addons.sponsorblock
+        pkgs.firefox-addons.tetrio-plus
+        pkgs.firefox-addons.translate-web-pages
+        pkgs.firefox-addons.tree-style-tab
+        pkgs.firefox-addons.tst-wheel-and-double
+        pkgs.firefox-addons.web-archives
+      ];
+
+      # Disable tab bar when using vertical tabs
+      userChrome = ''
+        #TabsToolbar { visibility: collapse !important; }
+      '';
+    };
+
+    programs.firefox.profiles.work = {
+      id = 2;
+
+      inherit (config.programs.firefox.profiles.base) search settings;
+
+      extensions = config.programs.firefox.profiles.base.extensions ++ [
+        pkgs.firefox-addons.multi-account-containers
+        pkgs.firefox-addons.open-url-in-container
+      ];
+    };
   };
 }