+++ modules/personal.nix
@@ -25,32 +25,6 @@
       always = true;
     }];
 
-    programs.firefox.profiles.default.extensions = [
-      pkgs.firefox-addons.bing-chat-for-all-browsers
-      pkgs.firefox-addons.copy-selected-links
-      pkgs.firefox-addons.ff2mpv
-      pkgs.firefox-addons.hover-zoom-plus
-      pkgs.firefox-addons.improved-tube
-      pkgs.firefox-addons.multi-account-containers
-      pkgs.firefox-addons.old-reddit-redirect
-      pkgs.firefox-addons.reddit-enhancement-suite
-      pkgs.firefox-addons.redirector
-      pkgs.firefox-addons.sponsorblock
-      pkgs.firefox-addons.tetrio-plus
-      pkgs.firefox-addons.translate-web-pages
-      pkgs.firefox-addons.tree-style-tab
-      pkgs.firefox-addons.tst-wheel-and-double
-      pkgs.firefox-addons.web-archives
-    ];
-
-    programs.firefox.profiles.work = {
-      id = 1;
-      inherit (config.programs.firefox.profiles.default) search settings;
-    };
-
-    # Disable tab bar when using vertical tabs
-    programs.firefox.profiles.default.userChrome = ''
-      #TabsToolbar { visibility: collapse !important; }
-    '';
+    programs.firefox.profiles.personal.isDefault = true;
   };
 }